+++ src/eslint-plugin-adeira/src/rules/EslintRule.flow.js
@@ -59,14 +59,14 @@ type Context = {|
 type ASTNodes = {|
   +'NewExpression:exit'?: (node: Node) => void,
   +'Property:exit'?: (node: Node) => void,
-  +BlockComment?: (node: Node) => void,
-  +CallExpression?: (node: Node) => void,
-  +Identifier?: (node: Node) => void,
-  +LineComment?: (node: Node) => void,
-  +NewExpression?: (node: Node) => void,
-  +Program?: (node: Node) => void,
-  +Property?: (node: Node) => void,
-  +ImportDeclaration?: (node: Node) => void,
+  +'BlockComment'?: (node: Node) => void,
+  +'CallExpression'?: (node: Node) => void,
+  +'Identifier'?: (node: Node) => void,
+  +'LineComment'?: (node: Node) => void,
+  +'NewExpression'?: (node: Node) => void,
+  +'Program'?: (node: Node) => void,
+  +'Property'?: (node: Node) => void,
+  +'ImportDeclaration'?: (node: Node) => void,
 |};
 
 export type EslintRule = {|